requirements.txt
fastapi==0.110.2
uvicorn==0.30.0
pytest==8.3.1
httpx==0.27.0
requests==2.32.5
grpcio>=1.80,<2
grpcio-tools>=1.80,<2
PyYAML==6.0.2
